+++ typademic/uploads/routes.py
@@ -4,55 +4,67 @@ import uuid
 from flask import session, render_template, request, send_file, redirect, url_for, current_app
 from sh import pandoc
 
-from typademic import limiter
+from typademic.app import limiter
 from typademic.uploads import blueprint
+from typademic.utils import remove_all_files_recursively
+
+@blueprint.route('/', methods=['GET'])
+def index():
+    if 'uid' not in session:
+        return render_template('index.html',
+                                   files=None,
+                                   error='')
+    else:
+        session_path = os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
+        files = os.listdir(session_path)
+        return render_template('index.html',
+                               files=files,
+                               error='')
 
 
-def uploaded_files():
-    try:
-        return os.listdir(os.path.join(current_app.config['UPLOADED_PATH'], session['uid']))
-    except Exception:
-        return None
-
-
-@blueprint.route('/', methods=['GET', 'POST'])
+@blueprint.route('/', methods=['POST'])
 def upload():
     if 'uid' not in session:
-        uid = uuid.uuid4().hex
-        session['uid'] = uid
-        session_upload_path = os.path.join(current_app.config['UPLOADED_PATH'], uid)
         try:
+            uid = uuid.uuid4().hex
+            session['uid'] = uid
+            session_path = os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
             # ensure the upload folder exists
-            os.mkdir(session_upload_path)
+            os.mkdir(session_path)
         except Exception as e:
             return render_template('index.html',
-                                   files=uploaded_files(),
+                                   files=None,
                                    error=str(e))
-    if request.method is 'POST':
-        try:
-            f = request.files.get('file')
-            f.save(os.path.join(current_app.config['UPLOADED_PATH'], session['uid'], f.filename))
-            return render_template('index.html',
-                                   files=uploaded_files(),
-                                   error='')
-        except Exception as e:
-            return render_template('index.html',
-                                   files=uploaded_files(),
-                                   error=str(e))
-    return render_template('index.html',
-                           files=uploaded_files(),
-                           error='')
+    session_path = os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
+
+    try:
+        f = request.files.get('file')
+        f.save(os.path.join(session_path, f.filename))
+        files = os.listdir(session_path)
+        return render_template('index.html',
+                               files=files,
+                               error='')
+    except Exception as e:
+        files = os.listdir(session_path)
+        return render_template('index.html',
+                                files=files,
+                               error=str(e))
 
 
 @blueprint.route('/clear', methods=['GET'])
 def clear():
-    try:
-        remove_all_recursively(os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
-        return redirect(url_for('upload'))
-    except Exception as e:
-        return render_template('index.html',
-                               files=uploaded_files(),
-                               error=str(e))
+    if 'uid' not in session:
+        return redirect(url_for('uploads.index'))
+    else:
+        session_path = os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
+        try:
+            remove_all_files_recursively(session_path)
+            return redirect(url_for('uploads.index'))
+        except Exception as e:
+            files = os.listdir(session_path)
+            return render_template('index.html',
+                                   files=files,
+                                   error=str(e))
 
 
 @blueprint.route('/clear_all/<key>', methods=['GET'])
@@ -60,42 +72,31 @@ def clear():
 def clear_all(key):
     if key is current_app.config['SECRET_KEY']:
         try:
-            remove_all_recursively(os.path.abspath(current_app.config['UPLOADED_PATH']))
+            remove_all_files_recursively(os.path.abspath(current_app.config['UPLOADED_PATH']))
             return render_template('index.html',
-                                   files=uploaded_files(),
+                                   files=None,
                                    error='All files are successfully removed.')
         except Exception as e:
             return render_template('index.html',
-                                   files=uploaded_files(),
+                                   files=None,
                                    error=str(e))
     else:
-        return redirect(url_for('upload'))
-
-
-def remove_all_recursively(path):
-    try:
-        for root, dirs, files in os.walk(path, topdown=False):
-            for name in files:
-                os.remove(os.path.join(root, name))
-            for name in dirs:
-                os.remove(os.path.join(root, name))
-        return None
-    except Exception as e:
-        return e
+        return redirect(url_for('uploads.index'))
 
 
 @blueprint.route('/render/<output_format>', methods=['GET'])
 def render(output_format):
-    if output_format not in ['docx', 'pdf']:
-        return redirect(url_for('upload'))
+    if output_format not in ['docx', 'pdf'] or 'uid' not in session:
+        return redirect(url_for('uploads.index'))
+    session_path = os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
+    files = os.listdir(session_path)
     output_filename = 'typademic.' + output_format
-    files = uploaded_files()
     try:
         md_files = ''
         for file in files:
             # Serve from cache
             if file.endswith(output_filename):
-                return send_file(os.path.join(current_app.config['UPLOADED_PATH'], session['uid'], output_filename),
+                return send_file(os.path.join(session_path, output_filename),
                                  attachment_filename=output_filename)
             # Extract md file(s)
             if file.endswith('.md'):
@@ -104,20 +105,19 @@ def render(output_format):
             return render_template('index.html',
                                    files=files,
                                    error='No Markdown file was uploaded. Please reset and try again.')
-        cwd = os.path.join(current_app.config['UPLOADED_PATH'], session['uid'])
         pandoc(md_files.strip(),
                '--output',
                output_filename,
                '--from',
                'markdown+ascii_identifiers+tex_math_single_backslash+raw_tex+table_captions+yaml_metadata_block+autolink_bare_uris',
                '--latex-engine=xelatex',
-               # Pandoc 2.2.3 fix
+               # FIXME Pandoc 2.2.3 fix
                # '--pdf-engine=xelatex',
                '--filter',
                'pandoc-citeproc',
                '--standalone',
-               _cwd=cwd)
-        return send_file(os.path.join(current_app.config['UPLOADED_PATH'], session['uid'], output_filename),
+               _cwd=session_path)
+        return send_file(os.path.join(session_path, output_filename),
                          attachment_filename=output_filename)
     except Exception as e:
         return render_template('index.html',
