+++ typademic/utils.py
@@ -1,0 +1,12 @@
+import os
+
+def remove_all_files_recursively(path):
+    try:
+        for root, dirs, files in os.walk(path, topdown=False):
+            for name in files:
+                os.remove(os.path.join(root, name))
+            for name in dirs:
+                os.remove(os.path.join(root, name))
+        return None
+    except Exception as e:
+        return e
